Restructure dockerfiles
Now, the test.Dockerfile is split in 4 different dockerfiles, 1 per type of test.
Despite the previous version being more convenient to work with, executing the
tests sequentially posed a huge toll on the time required to obtain the results.

With the new Dockerfiles, testing load can be parallelized between different containers.
This modification also implied the need for the evt-blind and edg-blind containers to receive
the baseline traces as part of their context. For this reason, the .dockerignore file has been
completely removed and the single Dockerfiles have been updated accordingly.